interfaces: move base declaration to the policy sub-package
The base declaration is becoming de-centralized and as a part of that it
will be composed of out of small pieces.

To make this possible the policy package needs to be able to enumerate
interfaces reliably and the best way to make that possible is to move it
to a separate pacakge and then call the public builtin.Interfaces
function.
