+++ interfaces/policy/basedeclaration_test.go
@@ -17,7 +17,7 @@
  *
  */
 
-package builtin_test
+package policy_test
 
 import (
 	"fmt"
